update: Translate variables to PT/BR
Change the input file from input.json to dados.json
and update the README accordingly. Modify the output
filename to formulario_preenchido.pdf for clarity.

Also, improved the script to accommodate
the changes in data structure and names in Portuguese.
This enhances usability and aligns with the
localization of the application.

Signed-off-by: Denis Yuji Shimizu <denis.shimizu@collabora.com>